Move download module out of setup
