+++ roms_tools/tests/test_setup/test_tides.py
@@ -1,7 +1,7 @@
 import pytest
 from roms_tools import Grid, TidalForcing
 import xarray as xr
-from roms_tools.setup.download import download_test_data
+from roms_tools.download import download_test_data
 import textwrap
 from pathlib import Path
 from conftest import calculate_file_hash
