+++ roms_tools/tests/test_setup/test_topography.py
@@ -1,7 +1,7 @@
 import pytest
 from roms_tools import Grid
 from roms_tools.setup.topography import _compute_rfactor
-from roms_tools.setup.download import download_test_data
+from roms_tools.download import download_test_data
 import numpy as np
 import numpy.testing as npt
 from scipy.ndimage import label
