+++ roms_tools/tests/test_setup/test_utils.py
@@ -1,6 +1,6 @@
 from roms_tools.setup.utils import interpolate_from_climatology
 from roms_tools.setup.datasets import ERA5Correction
-from roms_tools.setup.download import download_test_data
+from roms_tools.download import download_test_data
 import xarray as xr
